Replace types
